+++ wtpython/backends/__init__.py
@@ -1,0 +1,3 @@
+from .search_engine import SearchEngine  # noqa: F401
+from .stackoverflow import StackOverflow  # noqa: F401
+from .trace import Trace  # noqa: F401
