+++ wtpython/displays/__init__.py
@@ -1,0 +1,2 @@
+from .no_display import dump_info  # noqa: F401
+from .textual_display import TextualDisplay  # noqa: F401
